refactor: streamline chapter frontend internals
